ipcache: Split into multiple files
